Moved columnar format code into separate dir
