+++ cpp/src/arrow/colfmt/api.h
